requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
dataclasses==0.6
diff-match-patch==20241021
Django==5.1.4
django-admin-tools==0.9.3
django-ckeditor==6.7.2
django-import-export==4.3.7
django-js-asset==3.0.1
djangorestframework==3.15.2
djongo==1.2.31
dnspython==2.7.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.1
parse_pip_search==0.0.1
pillow==11.1.0
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.11
python-dotenv==1.0.1
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
soupsieve==2.6
sqlparse==0.5.3
tablib==3.8.0
tzdata==2024.2
urllib3==2.3.0
gunicorn

